refacor: remove main test file in project root
