+++ go.mod
@@ -4,6 +4,7 @@ go 1.24.2
 
 require (
 	github.com/golang-jwt/jwt/v5 v5.2.1
+	github.com/google/uuid v1.6.0
 	github.com/lib/pq v1.10.9
 	github.com/pressly/goose/v3 v3.21.1
 	github.com/redis/go-redis/v9 v9.5.3
