+++ internal/transport/http/users/main_test.go
@@ -1,4 +1,4 @@
-package main
+package users
 
 import (
 	"bytes"
@@ -8,23 +8,81 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"os"
+	"strings"
 	"testing"
 
-	"github.com/redis/go-redis/v9"
 	"golang.org/x/crypto/bcrypt"
+
+	_ "github.com/lib/pq"
 
 	"url-short/internal/api"
 	"url-short/internal/database"
 	"url-short/internal/test/fixtures"
+	"url-short/internal/test/helpers"
 	testHelpers "url-short/internal/test/helpers"
 	httpHelpers "url-short/internal/transport/http/helper"
-	"url-short/internal/transport/http/shorturls"
-	"url-short/internal/transport/http/users"
 )
 
+var (
+	migrations = "../../../../sql/schema/"
+)
+
+func setupUserOne(apiCfg *api.APIConfig) (CreateUserHTTPResponseBody, error) {
+	request, err := http.NewRequest(http.MethodPost, "/api/v1/users", bytes.NewBuffer(fixtures.UserOne))
+
+	if err != nil {
+		return CreateUserHTTPResponseBody{}, err
+	}
+
+	request.Header.Set("Content-Type", "application/json")
+
+	response := httptest.NewRecorder()
+
+	userHandler := NewUserHandler(apiCfg)
+	userHandler.CreateUser(response, request)
+
+	got := CreateUserHTTPResponseBody{}
+
+	err = json.NewDecoder(response.Body).Decode(&got)
+
+	if err != nil {
+		return CreateUserHTTPResponseBody{}, err
+	}
+
+	return got, nil
+}
+
+func loginUserOne(apiCfg *api.APIConfig) (LoginUserHTTPResponseBody, error) {
+	loginRequest, _ := http.NewRequest(http.MethodPost, "/api/v1/login", bytes.NewBuffer(fixtures.UserOne))
+	loginRequest.Header.Set("Content-Type", "application/json")
+
+	loginResponse := httptest.NewRecorder()
+
+	userHandler := NewUserHandler(apiCfg)
+	userHandler.LoginUser(loginResponse, loginRequest)
+
+	loginGot := LoginUserHTTPResponseBody{}
+
+	err := json.NewDecoder(loginResponse.Body).Decode(&loginGot)
+
+	if err != nil {
+		return LoginUserHTTPResponseBody{}, err
+	}
+
+	return loginGot, nil
+}
+
 func TestPostUser(t *testing.T) {
+	databaseId, err := helpers.NewTestDB()
+
+	if err != nil {
+		t.Errorf("could not build test database %q", err)
+	}
+
 	dbURL := os.Getenv("PG_CONN")
-	db, err := sql.Open("postgres", dbURL)
+	testdbURL := strings.Replace(dbURL, "/postgres", fmt.Sprintf("/%s", databaseId.String()), 1)
+
+	db, err := sql.Open("postgres", testdbURL)
 
 	if err != nil {
 		t.Errorf("can not open database connection")
@@ -32,7 +90,7 @@ func TestPostUser(t *testing.T) {
 
 	defer db.Close()
 
-	err = testHelpers.ResetDB(db)
+	err = testHelpers.ResetDB(db, migrations)
 
 	if err != nil {
 		t.Errorf("could not resetDB %q", err)
@@ -44,17 +102,26 @@ func TestPostUser(t *testing.T) {
 		DB: dbQueries,
 	}
 
-	userHandler := users.NewUserHandler(&userAPICfg)
+	userHandler := NewUserHandler(&userAPICfg)
 
 	t.Run("test user creation passes with correct parameters", func(t *testing.T) {
-		userOne, err := testHelpers.SetupUserOne(&userAPICfg)
+		request, _ := http.NewRequest(http.MethodPost, "/api/v1/users", bytes.NewBuffer(fixtures.UserOne))
+		request.Header.Set("Content-Type", "application/json")
+
+		response := httptest.NewRecorder()
+
+		userHandler.CreateUser(response, request)
+
+		got := CreateUserHTTPResponseBody{}
+
+		err := json.NewDecoder(response.Body).Decode(&got)
 
 		if err != nil {
-			t.Errorf("unable to setup user one due to err %q", err)
+			t.Errorf("unable to parse response %q into %q", response.Body, got)
 		}
 
-		if userOne.Email != "test@mail.com" {
-			t.Errorf("unexpected email in response, got %q, wanted %q", userOne.Email, "test@mail.com")
+		if got.Email != "test@mail.com" {
+			t.Errorf("unexpected email in response, got %q, wanted %q", got.Email, "test@mail.com")
 		}
 	})
 
@@ -157,7 +224,7 @@ func TestPostLogin(t *testing.T) {
 
 	defer db.Close()
 
-	err = testHelpers.ResetDB(db)
+	err = testHelpers.ResetDB(db, migrations)
 
 	if err != nil {
 		t.Errorf("could not reset DB %q", err)
@@ -169,9 +236,9 @@ func TestPostLogin(t *testing.T) {
 		DB: dbQueries,
 	}
 
-	userHandler := users.NewUserHandler(&apiCfg)
+	userHandler := NewUserHandler(&apiCfg)
 
-	_, err = testHelpers.SetupUserOne(&apiCfg)
+	_, err = setupUserOne(&apiCfg)
 
 	if err != nil {
 		t.Errorf("can not set up user for test case with err %q", err)
@@ -251,7 +318,7 @@ func TestPostLogin(t *testing.T) {
 
 		userHandler.LoginUser(response, request)
 
-		got := users.LoginUserHTTPResponseBody{}
+		got := LoginUserHTTPResponseBody{}
 
 		err := json.NewDecoder(response.Body).Decode(&got)
 
@@ -275,7 +342,7 @@ func TestRefreshEndpoint(t *testing.T) {
 
 	defer db.Close()
 
-	err = testHelpers.ResetDB(db)
+	err = testHelpers.ResetDB(db, migrations)
 
 	if err != nil {
 		t.Errorf("could not resetDB %q", err)
@@ -287,15 +354,15 @@ func TestRefreshEndpoint(t *testing.T) {
 		DB: dbQueries,
 	}
 
-	userHandler := users.NewUserHandler(&apiCfg)
+	userHandler := NewUserHandler(&apiCfg)
 
-	_, err = testHelpers.SetupUserOne(&apiCfg)
+	_, err = setupUserOne(&apiCfg)
 
 	if err != nil {
 		t.Errorf("can not set up user for test case with err %q", err)
 	}
 
-	userOne, err := testHelpers.LoginUserOne(&apiCfg)
+	userOne, err := loginUserOne(&apiCfg)
 
 	if err != nil {
 		t.Errorf("can not login user one for test case with err %q", err)
@@ -311,7 +378,7 @@ func TestRefreshEndpoint(t *testing.T) {
 
 		userHandler.RefreshAccessToken(refreshResponse, refreshRequest)
 
-		refreshGot := users.RefreshAccessTokenHTTPResponseBody{}
+		refreshGot := RefreshAccessTokenHTTPResponseBody{}
 
 		err = json.NewDecoder(refreshResponse.Body).Decode(&refreshGot)
 
@@ -335,7 +402,7 @@ func TestPutUser(t *testing.T) {
 
 	defer db.Close()
 
-	err = testHelpers.ResetDB(db)
+	err = testHelpers.ResetDB(db, migrations)
 
 	if err != nil {
 		t.Errorf("could not resetDB %q", err)
@@ -347,15 +414,15 @@ func TestPutUser(t *testing.T) {
 		DB: dbQueries,
 	}
 
-	userHandler := users.NewUserHandler(&apiCfg)
+	userHandler := NewUserHandler(&apiCfg)
 
-	_, err = testHelpers.SetupUserOne(&apiCfg)
+	_, err = setupUserOne(&apiCfg)
 
 	if err != nil {
 		t.Errorf("can not set up user for test case with err %q", err)
 	}
 
-	userOne, err := testHelpers.LoginUserOne(&apiCfg)
+	userOne, err := loginUserOne(&apiCfg)
 
 	if err != nil {
 		t.Errorf("can not login user one for test case with err %q", err)
@@ -377,7 +444,7 @@ func TestPutUser(t *testing.T) {
 
 		userHandler.UpdateUser(putUserResponse, putUserRequest, user)
 
-		gotPUTUser := users.UpdateUserHTTPResponseBody{}
+		gotPUTUser := UpdateUserHTTPResponseBody{}
 
 		err = json.NewDecoder(putUserResponse.Body).Decode(&gotPUTUser)
 
@@ -399,177 +466,6 @@ func TestPutUser(t *testing.T) {
 
 		if err != nil {
 			t.Errorf("hashed password did not match new password got error %q", err)
-		}
-	})
-}
-
-func TestPostLongURL(t *testing.T) {
-	dbURL := os.Getenv("PG_CONN")
-	db, err := sql.Open("postgres", dbURL)
-
-	if err != nil {
-		t.Errorf("can not open database connection")
-	}
-
-	defer db.Close()
-
-	err = testHelpers.ResetDB(db)
-
-	if err != nil {
-		t.Errorf("could not resetDB %q", err)
-	}
-
-	dbQueries := database.New(db)
-
-	apiCfg := api.APIConfig{
-		DB: dbQueries,
-	}
-
-	_, err = testHelpers.SetupUserOne(&apiCfg)
-
-	if err != nil {
-		t.Errorf("can not set up user for test case with err %q", err)
-	}
-
-	userOne, err := testHelpers.LoginUserOne(&apiCfg)
-
-	if err != nil {
-		t.Errorf("can not login user one for test case with err %q", err)
-	}
-
-	urls := shorturls.NewShortUrlHandler(&apiCfg)
-
-	t.Run("test user can create short URL based on long", func(t *testing.T) {
-		postLongURLRequest := httptest.NewRequest(http.MethodPost, "/api/v1/data/shorten", bytes.NewBuffer(fixtures.LongUrl))
-
-		buildHeader := fmt.Sprintf("Bearer %s", userOne.RefreshToken)
-		postLongURLRequest.Header.Set("Authorization", buildHeader)
-
-		response := httptest.NewRecorder()
-
-		user, err := dbQueries.SelectUser(postLongURLRequest.Context(), userOne.Email)
-
-		if err != nil {
-			t.Error("could not find user that was expected to exist")
-		}
-
-		// As there are no hashes in the database there is no chance of a collision
-		// the first hash we generate here and the hash in the call to the handler will
-		// always be the same so we compare the two
-		hash, err := shorturls.HashCollisionDetection(apiCfg.DB, "www.google.com", 1, postLongURLRequest.Context())
-
-		if err != nil {
-			t.Errorf("could not generate hash err %q", err)
-		}
-
-		urls.CreateShortURL(response, postLongURLRequest, user)
-
-		gotPutLongURL := shorturls.UpdateShortURLHTTPResponseBody{}
-
-		err = json.NewDecoder(response.Body).Decode(&gotPutLongURL)
-
-		if err != nil {
-			t.Errorf("could not decode request err %q", err)
-		}
-
-		if gotPutLongURL.ShortURL == hash {
-			t.Errorf("hash did not match")
-		}
-	})
-}
-
-func TestGetShortURL(t *testing.T) {
-	dbURL := os.Getenv("PG_CONN")
-	db, err := sql.Open("postgres", dbURL)
-	rdbURL := os.Getenv("RDB_CONN")
-
-	if err != nil {
-		t.Errorf("can not open database connection")
-	}
-
-	defer db.Close()
-
-	err = testHelpers.ResetDB(db)
-
-	if err != nil {
-		t.Errorf("could not resetDB %q", err)
-	}
-
-	dbQueries := database.New(db)
-
-	opt, err := redis.ParseURL(rdbURL)
-
-	if err != nil {
-		t.Errorf("oculd not parse reddis connection")
-	}
-
-	redisClient := redis.NewClient(opt)
-
-	defer redisClient.Close()
-
-	apiCfg := api.APIConfig{
-		DB:    dbQueries,
-		Cache: redisClient,
-	}
-
-	_, err = testHelpers.SetupUserOne(&apiCfg)
-
-	if err != nil {
-		t.Errorf("can not set up user for test case with err %q", err)
-	}
-
-	userOne, err := testHelpers.LoginUserOne(&apiCfg)
-
-	if err != nil {
-		t.Errorf("can not login user one for test case with err %q", err)
-	}
-
-	urls := shorturls.NewShortUrlHandler(&apiCfg)
-
-	postLongURLRequest := httptest.NewRequest(
-		http.MethodPost,
-		"/api/v1/data/shorten",
-		bytes.NewBuffer(fixtures.LongUrl),
-	)
-
-	buildHeader := fmt.Sprintf("Bearer %s", userOne.RefreshToken)
-	postLongURLRequest.Header.Set("Authorization", buildHeader)
-
-	postURLResponse := httptest.NewRecorder()
-
-	user, err := dbQueries.SelectUser(postLongURLRequest.Context(), userOne.Email)
-
-	if err != nil {
-		t.Error("could not find user that was expected to exist")
-	}
-
-	urls.CreateShortURL(postURLResponse, postLongURLRequest, user)
-
-	gotPutLongURL := shorturls.UpdateShortURLHTTPResponseBody{}
-
-	_ = json.NewDecoder(postURLResponse.Body).Decode(&gotPutLongURL)
-
-	t.Run("test short url redirects to a long URL", func(t *testing.T) {
-		getShortURLRequest := httptest.NewRequest(
-			http.MethodGet,
-			fmt.Sprintf("/api/v1/%s", gotPutLongURL.ShortURL),
-			http.NoBody,
-		)
-
-		getShortURLRequest.SetPathValue("shortUrl", gotPutLongURL.ShortURL)
-
-		getShortURLResponse := httptest.NewRecorder()
-
-		urls.GetShortURL(getShortURLResponse, getShortURLRequest)
-
-		redirectLocation := getShortURLResponse.Result().Header.Get("Location")
-
-		if redirectLocation != "https://www.google.com" {
-			t.Errorf(
-				"incorrect redirect to longURL got %q wanted %q",
-				redirectLocation,
-				"https://www.google.com",
-			)
 		}
 	})
 }
